krux_finance_ocr-main/requirements.txt
transformers
datasets
torch
torchvision
pillow
pytesseract
pdf2image
accelerate
scikit-learn
numpy
pandas
fastapi
uvicorn
python-multipart
